# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk Sistem Data Induk Siswa SMK",
        "title": "API Siswa Induk",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@siswa.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/prestasi/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete achievement record",
                "tags": [
                    "Prestasi"
                ],
                "summary": "Delete achievement",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Achievement Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/siswa/{id}/prestasi": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add achievement record to student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prestasi"
                ],
                "summary": "Add achievement",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Achievement data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreatePrestasiRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.PrestasiResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.CreatePrestasiRequest": {
            "type": "object",
            "required": [
                "bidang"
            ],
            "properties": {
                "bidang": {
                    "type": "string",
                    "enum": [
                        "Kesenian",
                        "Olahraga",
                        "Kemasyarakatan",
                        "Pramuka",
                        "Karya Tulis",
                        "Lainnya"
                    ],
                    "example": "Olahraga"
                },
                "keterangan": {
                    "type": "string",
                    "example": "Juara 1 Lomba Lari"
                },
                "tahun": {
                    "type": "integer",
                    "example": 2024
                },
                "tingkat": {
                    "type": "string",
                    "enum": [
                        "Sekolah",
                        "Kecamatan",
                        "Kota",
                        "Provinsi",
                        "Nasional",
                        "Internasional"
                    ],
                    "example": "Kota"
                }
            }
        },
...
        "responses.PrestasiResponse": {
            "type": "object",
            "properties": {
                "bidang": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keterangan": {
                    "type": "string"
                },
                "tahun": {
                    "type": "integer"
                },
                "tingkat": {
                    "type": "string"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\" to authenticate",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}